Separate the generic config panel in 2 classes.
One general, and one for detectors, with the preview button panel.
